Move handler controllers to an internal package.
