--- engine/controller/doc.go
@@ -1,3 +1,0 @@
-// Package controller contains interfaces that the engine uses to consume
-// controllers for each of the message handler types.
-package controller
