+++ engine/internal/aggregate/doc.go
