+++ engine/internal/aggregate/scope.go
