+++ engine/internal/aggregate/scope_test.go
@@ -8,9 +8,9 @@ import (
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/dogma/fixtures"
-	. "github.com/dogmatiq/testkit/engine/controller/aggregate"
 	"github.com/dogmatiq/testkit/engine/envelope"
 	"github.com/dogmatiq/testkit/engine/fact"
+	. "github.com/dogmatiq/testkit/engine/internal/aggregate"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	. "github.com/onsi/gomega/gstruct"
@@ -346,7 +346,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/aggregate/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/aggregate/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -385,7 +385,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/aggregate/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/aggregate/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
