+++ engine/internal/process/scope_test.go
@@ -8,9 +8,9 @@ import (
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/dogma/fixtures"
-	. "github.com/dogmatiq/testkit/engine/controller/process"
 	"github.com/dogmatiq/testkit/engine/envelope"
 	"github.com/dogmatiq/testkit/engine/fact"
+	. "github.com/dogmatiq/testkit/engine/internal/process"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	. "github.com/onsi/gomega/gstruct"
@@ -115,7 +115,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -206,7 +206,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -246,7 +246,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -286,7 +286,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -522,7 +522,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -562,7 +562,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -652,7 +652,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
@@ -695,7 +695,7 @@ var _ = Describe("type scope", func() {
 							"Location": MatchAllFields(
 								Fields{
 									"Func": Not(BeEmpty()),
-									"File": HaveSuffix("/engine/controller/process/scope_test.go"),
+									"File": HaveSuffix("/engine/internal/process/scope_test.go"),
 									"Line": Not(BeZero()),
 								},
 							),
