+++ engine/internal/projection/scope.go
