+++ engine/internal/projection/scope_test.go
@@ -7,9 +7,9 @@ import (
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/dogma/fixtures"
-	. "github.com/dogmatiq/testkit/engine/controller/projection"
 	"github.com/dogmatiq/testkit/engine/envelope"
 	"github.com/dogmatiq/testkit/engine/fact"
+	. "github.com/dogmatiq/testkit/engine/internal/projection"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
